feat(toolbar): add default toolbar component (#266)
The changes introduce a new `DefaultToolbar` component that can be used in place of the `TransparentToolbar` component. The `DefaultToolbar` component includes an optional `title` prop that can be used to display a title within the toolbar.

The changes also update the `Toolbar` component to use the `DefaultToolbar` component when the current route matches the "Plans" route.