Left sidebar sorting (#899)
* feat(notes-list): Sort sessions by event start date or creation date

The changes in this commit focus on improving the sorting of sessions in the notes list component. Previously, the sessions were sorted solely based on their creation date, but now they are sorted based on the start date of the associated event, if available. If the event start date is not available, the creation date of the session is used as a fallback.

This change ensures that the sessions are displayed in the correct chronological order, taking into account the actual event dates rather than just the session creation dates.

* i18n

* dirty fix

---------

Co-authored-by: yujonglee <yujonglee.dev@gmail.com>